refactor(ruleset): add builtins/geo dir
