+++ ruleset/src/builtins/geo/v2geo/mod.rs
@@ -2,5 +2,3 @@
 
 pub mod load;
 pub mod v2geo;
-
-pub use v2geo as geo;
