+++ ruleset/src/lib.rs
@@ -5,8 +5,8 @@ use std::fmt::{self, Debug};
 use std::net::IpAddr;
 use std::sync::Arc;
 
+pub mod builtins;
 pub mod expr_rule;
-pub mod v2geo;
 
 #[derive(Debug)]
 pub enum Protocol {
